# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Unified Ads Management SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /platforms:
    get:
      operationId: listPlatforms
      tags: [platforms]
      responses:
        "200":
          description: Connected platforms
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Platform" }
    post:
      operationId: connectPlatform
      tags: [platforms]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConnectPlatformRequest" }
      responses:
        "200":
          description: Connected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Platform" }
  /platforms/{platformId}:
    delete:
      operationId: disconnectPlatform
      tags: [platforms]
      parameters:
        - in: path
          name: platformId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /platforms/{platformId}/test:
    post:
      operationId: testPlatformConnection
      tags: [platforms]
      parameters:
        - in: path
          name: platformId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformTestResult" }
...
  /publish:
    post:
      operationId: publishCampaign
      tags: [publish]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublishRequest" }
      responses:
        "200":
          description: Publish jobs queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublishResponse" }
  /publish/{jobId}/retry:
    post:
      operationId: retryPublishJob
      tags: [publish]
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Retried
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublishJob" }
  /status:
    get:
      operationId: getPublishStatus
      tags: [publish]
      parameters:
        - in: query
          name: campaignId
          schema: { type: string }
      responses:
        "200":
          description: Publish job status across platforms
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PublishJob" }
  /media:
    get:
      operationId: listMedia
      tags: [media]
      responses:
        "200":
          description: Media library
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MediaAsset" }
    post:
      operationId: createMedia
      tags: [media]
      description: Register a media asset. URL-based upload (would be R2 in production).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMediaRequest" }
      responses:
        "200":
          description: Media created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaAsset" }
  /media/{mediaId}:
    delete:
      operationId: deleteMedia
      tags: [media]
      parameters:
        - in: path
          name: mediaId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
components:
  schemas:
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    Platform:
      type: object
      properties:
        id: { type: string }
        key: { type: string }
        name: { type: string }
        accountLabel: { type: string }
        status: { type: string, enum: [active, expired, error] }
        connectedAt: { type: string, format: date-time }
        lastSyncedAt: { type: string, format: date-time, nullable: true }
        config:
          type: object
          additionalProperties: true
      required: [id, key, name, accountLabel, status, connectedAt]
    ConnectPlatformRequest:
      type: object
      properties:
        key: { type: string }
        accountLabel: { type: string }
        apiKey: { type: string, nullable: true }
        config:
          type: object
          additionalProperties: true
      required: [key, accountLabel]
    PlatformTestResult:
      type: object
      properties:
        ok: { type: boolean }
        latencyMs: { type: integer }
        message: { type: string }
      required: [ok, latencyMs, message]
...
    PublishRequest:
      type: object
      properties:
        campaignId: { type: string }
        platforms:
          type: array
          items: { type: string }
      required: [campaignId]
    PublishResponse:
      type: object
      properties:
        campaignId: { type: string }
        jobs:
          type: array
          items: { $ref: "#/components/schemas/PublishJob" }
      required: [campaignId, jobs]
    PublishJob:
      type: object
      properties:
        id: { type: string }
        campaignId: { type: string }
        platformKey: { type: string }
        platformName: { type: string }
        status: { type: string, enum: [pending, processing, success, failed, retrying] }
        attempts: { type: integer }
        externalId: { type: string, nullable: true }
        externalUrl: { type: string, nullable: true }
        error: { type: string, nullable: true }
        response:
          type: object
          additionalProperties: true
          nullable: true
        impressions: { type: integer }
        clicks: { type: integer }
        spend: { type: number }
        startedAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, campaignId, platformKey, platformName, status, attempts, impressions, clicks, spend, createdAt]
    MediaAsset:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [image, video] }
        name: { type: string }
        url: { type: string }
        sizeBytes: { type: integer }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, type, name, url, sizeBytes, createdAt]
    CreateMediaRequest:
      type: object
      properties:
        type: { type: string, enum: [image, video] }
        name: { type: string }
        url: { type: string }
        sizeBytes: { type: integer }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
      required: [type, name, url]
...